Split out storage files for clarity
The problem with reading through, or doing anything in storage, is that
there are giant files with no specific order. It's really hard to read
through and understand the different parts of storage. By splitting them
out it becomes more clear.
