+++ ironfish/src/storage/levelup/index.ts
@@ -4,8 +4,6 @@
 
 export * from './batch'
 export * from './database'
-export * from './encoding'
-export * from './errors'
 export * from './store'
 export * from './transaction'
-export * from './types'
+export * from './utils'
